Add files via upload
This script is a wrapper for Bedtools Multiinter that allows you to intersect multiple files in a strand aware manner. It won't take any options (although it could be easy to make a version with changed options) and it prints the file names in the header. Otherwise read the help (-h) for the script.